optimise config code
